Generally use reliable shutdown method and test shutdown applet separately.
Sometimes Sikuli fails to find the shutdown icon in the shutdown
applet's menu (possibly because it's covered by a notification) so
that method isn't completely reliable at the moment. To contain this
problem we test the shutdown applet separately, and rely on the more
dependable remote shell for halting in the general case.

While we're at it we also make sure that the shutdown steps are
formulated so what they do is more explicit.