Prevent iproute2 from being installed from wheezy-backports (Closes: #7337).
This package has "Priority: important", so it's automatically installed by
tasksel or similar early during the Tails/Wheezy boot process. But it conflicts
with Wheezy's iproute, that is itself installed later, and then iproute2 is
deinstalled, but not purged. To fix that, and make the build process clearer,
let's prevent iproute2 from being installed from wheezy-backports entirely.

Note that iproute2 is not part of Wheezy. The chosen pinning section only
affects wheezy-backports, so that it should be innocuous when we start migrating
to Jessie.