Drop all search engine re-configuration for the Unsafe Browser.
The Search Bar is removed in the Windows 8 camouflage, so having
Wikipedia as the default cannot be used as a stealthy visual indicator
any more. We already have "Unsafe Browser" in both the window title,
and the button to the right of the URL box, which should make it clear
enough which browser is used.