Include syslinux.exe for win32 in utils/win32/ on the ISO filesystem.
The APT config mangling is made in the chroot, as that's where all keys
corresponding to the configured APT repos can be found, and the APT pinning has
been setup. Note that, when this hook runs, the SquashFS has been generated
already, so modifying the content of the chroot at this point should not affect
the resulting ISO much.