Announce no upgrade paths at all from 1.1~rc1 to 1.1.
This reverts commit 8c1dbf1535dbbeb1451ad076a93919517ea36793 and part
of 421a0973f6dd6616514567c983a67c00481a44a8.

Due to #7641, this incremental upgrade is broken, so we've documented
a workaround in the release notes, and rely on the fact that users will see it
on the news page, when starting the browser.

Other rejected "solutions":

* if we kept announcing that upgrade path, then users would see the IUK
  downloaded, and then an error message, which is a waste of time and plenty
  of confusion.
* if we only announced a full upgrade, then users would see it before reading
  the news, and be told that there's no incremental upgrade, and be pointed to
  a page that says the contrary. Confusing.

We also want to avoid concurrently running tails-upgrade-frontend's that display
conflicting messages, which this solution avoids.