Disable runtime APT proxy configuration when using APT in binary_local-hooks (Closes: #7691).
If we don't do this, APT can't download lists from the Internet. And then, in
some situations, it lacks the information needed to install dpkg-dev. I'm not
sure why it can even work in some other cases, but in any case, this hook
makes oh so much sense to me with this modification.