Install Plymouth via chroot_local-hooks.
In lb 2.x, the "standard" packages list pulls console-common in, which plymouth
conflicts with, so we have to deal with that at this stage.

This implies to patch the plymouth initscript via chroot_local-hooks too,
as chroot_local-patches are applied before we run our hooks, so before
this file is installed.