Fix localization of the Unsafe Browser's name-setting.
It turns out that the branding's name of the browser can differ quite
a bit between different langpacks, e.g. in Arabic it's written in
Arabic, but in languages using latin based it's often "(Mozilla)?
Firefox". Hence we must set the value according to the key, which is
the saner way, instead of simply replacing values.