Revert "$HOME(pidgin): hide formatting toolbar" (Closes: #7356)
This reverts commit 5bdf66181d704dc801a68c3f1ee1343cba8290f9.

Having the formatting toolbar displayed in Pidgin is the original default, and
it makes the OTR status more explicit by displaying it with words.

Without this toolbar, the OTR status have to be inferred either from the chat
log, or from the small logo on the right of the OTR menu.