Add localization helper functions for XUL applications.
That is for the various browsers and Tor Launcher. This will make it
possible to drop the browser localization code from Tails Greeter's
PostLogin script, where that really doesn't belong.

Also, this new code is an improvement in that it will pick a "similar"
locale if the specific one we look for isn't available. E.g. with
pt-BR we will get pt, and the Firefox is smart enough to use the pt-PT
langpack.