Use the correct stack of rootfs:s for the chroot browsers (Closes: #8152, #8152).
After installing incremental upgrades Tails' root filesystem consists
of a stack squashfs:s, not only filesystem.squashfs. When not stacking
them correct we may end up using the Tor Browser (Firefox) from an
older version of Tails, or with no Tor Browser at all, as in the
upgrade from Tails 1.1.2 to 1.2, when we migrated from Iceweasel to
the Tor Browser.

Based on patch contributed by sanic.