Have APT directly use the Tor SOCKS proxy (Closes: #8194).
At the end of the chroot_local-hooks stage, we replace the hook that used to
tell APT to use the Polipo HTTP proxy with another one that does (a more
elaborate version) s,http://,tor+http://, in APT sources.

Then, APT will use the "tor+http" method, that we install in
/usr/lib/apt/methods, and that is a simple torsocks wrapper for the
"http" method.