Adjust hook that includes syslinux in the ISO filesystem to APT being moved from HTTP to SOCKS.
Don't run apt-get update in the chroot again after reconfiguring APT sources
for runtime.

This is broken by the move to torsocks. Back when we used a HTTP proxy, the
HTTP_PROXY environment variable passed by the build environment was overriding
the APT proxy configured for runtime. This is not the case anymore, and now that
APT run tries to use both the tor+http shim and the HTTP proxy used in the build
environment, which is rejected by torsocks.

Besides, the build process works just fine without this "apt-get update" run,
so let's just drop it.