Don't look for the "Tor is ready" notification.
In Wheezy notifications already became unreliable too look for since
they always timeout after 5 seconds. Now in Jessie we do not only have
that problem, but the notification is placed on top of GNOME's bottom
panel *and* is transparent, so any application entries there will
affect how the notification appears.

The "Tor is ready" step already makes sure that Tor works by looking
at the exact same thing that triggers the notification, so we do not
lose anything. Well, except checking that the notification works
(which it didn't until the previous commit) but we should then test
that in its own step where we for instance make sure that the bottom
panel is empty, or similar.