Don't bother creating /var/lib/live in tails-detect-virtualization.
First, our live-config scripts write stuff in there, so if this directory
doesn't exist, we'll notice or we have bigger problems.

Second, even if this directory didn't exist or wasn't writable, then the
following line will fail, and then the unit will enter failed state, and we'll
notice. That's also why I didn't use ConditionPathIsDirectory= nor
ConditionPathIsReadWrite=, as these merely skip the unit (without putting it in
failed state) if the conditions are not met.