Repair emergency shutdown on boot medium removal (Closes: #8098).
The GDM unit file has a Restart=always directive, which is good in the general
case. However, it breaks our emergency shutdown on boot medium removal feature.
Let's disable it, then.

Also, the kexec-load initscript normally silently exits unless systemd is
currently running a reboot job. This is not the case when the emergency shutdown
has been triggered, so let's remove this check. (And besides, to be able to run
this check after the boot medium has been removed, we would have to lock way
more files in memory).