Stack another workaround on #8271.
Unfortunately the workaround in commit 992cc82 breaks the "Tails seems
to have booted normally" step if "GNOME has started" (with the
workaround) has run before. That breaks, for instance, the "Tails can
boot from live systems stored on hard drives" in
untrusted_partitions.feature.

Just to be extra clear: this definitely is a temporary workaround.