Don't try to switch to tty1 on emergency shutdown.
It is blocking emergency shutdown from a GNOME session (but not from GDM, for
some reason)... but manually switching with CTRL-ALT-F1 suddenly unblocks it.
My *guess* is that logind is trying to do some magics on VT switch, that fails
since we haven't locked the needed files into memory, and then it blocks chvt
from returning.

At least for the time being, let's simplify this part and just stay on the VT
we're on. We can attempt to make it nicer later on.