Don't patch the plymouth initscript anymore.
It's useless, since the plymouth package now ships native systemd unit files.
And then, keeping this patch around would be misleading.

The right way to disable the plymouth splash screen on shutdown would nowadays
be to remove the plymouth-poweroff.service symlink from
/lib/systemd/system/poweroff.target.wants/, and same for the reboot, halt, etc.
targets. However, I can't see it making a difference in my VM (where the screen
is garbled on regular shutdown anyway), so I'm not touching these bits yet.
And who knows, on other hardware, the shutdown splash screen might now be
working, and we could possibly use it to tell the user what's happening in
a nicer way.