Loosen a bit the AppArmor service's dependencies.
It currently blocks the boot for 10 seconds, since it's run early, but after
$remote_fs.

Let's start it with the default dependencies, that is as part of what's the
equivalent of runlevel 2 in systemd context. This way, it can be started in
parallel with other services.

Note that this works because we don't confine with AppArmor services that are
run before login, except cupd.service, for which this commit adds a dependency
on apparmor.service.