Torify Git with tsocks, instead of setting GIT_PROXY_COMMAND (Will-Fix: #8680).
GIT_PROXY_COMMAND is taken into account for git:// and Git over SSH, but not for
Git over HTTPS. The latter use to work when we were shipping Polipo, probably
thanks to the HTTP_PROXY environment variable and friends. Now it's broken.

Here, we're dropping usage of GIT_PROXY_COMMAND, and replacing it with a tsocks
wrapper.

Note: it would be nicer to use torsocks in this wrapper, which would work for
git:// and HTTPS, but it would break Git over SSH, since in Tails SSH has its
own ProxyCommand set to connect-socks, and then torsocks would block connections
initiated by connect-socks to the Tor SOCKS proxy, on the ground that they're
aimed at localhost and thus might be DNS requests.