Decouple VMStorage class from VM class.
When we splitted out VMNet from VM, we also decoupled the Libvirt
connection ($virt) and made it close after each feature. However, the
VMStorage instance was still running a single instance throughout the
feature so when a feature ended it would keep the reference to the old
now-closed Libvirt connection. In other words, VMStorage would only
work for the first feature.

Instead we create a new VMStorage instance per feature, just like with
VMNet.