WIP: Rename the Tor Browser "download" directory, again. And add automated tests for the Tor Browser.
The idea is to make sure that the AppArmor confinement doesn't break too much
functionality, and actually confines the browser a bit.

Sorry for the non-atomic commit: getting directory name changing under my feet
while the automated tests are being drafted made it too hard to split this in
a nice way.