Also install the syslinux 32-bit UEFI boot loader.
Note that we install it in BOOT/TAILS32 (while the 64-bit boot loader lives in
the BOOT/EFI fallback directory), since syslinux module files have the same name
regardless of the architecture, so we cannot simply put them all in the
same directory.

The drawback is that some buggy UEFI firmware may pick the wrong bootloader,
that's not made for their architecture. Or, that they present the user with
confusing choice between "EFI" and "TAILS32". We'll see.

Of course, alternatively we could try to rename the 32-bit module files, so we
can put them in BOOT/EFI as well; this would require adjusting the configuration
files accordingly, in order to 1. load the 32-bit modules when the UEFI firmware
is 32-bit; and 2. convince syslinux to load a different configuration file
depending on the UEFI firmware's architecture (*not* the CPU architecture,
beware!).

Refs: #8471