Reorder steps to avoid lost key presses.
First we start the Tor Browser, then we run the step 'there is a GNOME
bookmark for the amnesiac Tor Browser directory' which opens the GNOME
Places menu. It seems this may cause the Ctrl+S key presses in the
next step 'I can save the current page as "index.html" to the default
downloads directory' to not reach the Tor Browser, but probably get
lost in the Places menu before it closes from the Esc key press
(Sikuli is fast!), so the step fails. The new step ordering prevents
this from happening.