Zero-pad our ISO images so their sizes are divisible by 2048.
The data part of an ISO image's sectors are 2048 bytes, which implies
that ISO images should always have a size divisible by 2048. Some
application, e.g. VirtualBox, use this as a sanity check, treating ISO
images for which this isn't true as garbage.

Our isohybrid post-processing seems to result in ISO images with sizes
that are multiples of 1024, so we now pad our ISO images (if needed)
to ensure divisibility with 2048 to avoid issues with e.g. VirtualBox.

Will-fix: #8891