Import "clean" OpenPGP keys.
This will not affect the functionality of the existing steps, but it
will be useful for future tests, such as those for syncing keys using
Seahorse. (Seahorse doesn't seem provide any visual feedback when the
sync is successful, so instead we'll check a key's signature count
before and after sync).