Move to GitLab for the "fork us on.." documentation.
Its UI is way nicer than repo.or.cz, it is easier to configure, it offers HTTPS,
and most importantly we can easily push there ourselves (to get an up-to-date
mirror, and e.g. remove old tags or rewrite history) as opposed to our old
mirror-style repo on repo.or.cz.