Improve OTR-integration when sending messages.
We do this by adding the message that was so carefully crafted by
jabbebot to OTR's "opdata", and use that one as a basis. In the old
way we crafted a new message ourselves from scratch, but that made us
lose other message attributes, like the type, so groupchat messages
couldn't be sent properly.