Robustness fix.
In general, when opening menus and then clicking on one of its menu
entries, we should use wait_and_click() for at least the latter, since
there is a slight delay between when we click on the menu and when the
menu opens, but we'll start looking for the entry *immediately* after
clicking.

Thanks for identifying this instance of this problem, intrigeri!