Ensure the major KMS modules are included in the initramfs.
The goal is to have a more seamless transition from initial kernel modesetting
to X.Org when booting, and same on the way back on shutdown. It also will be
needed to get a nicer, Tails-specific boot splash. Moreover, as seen in the
bugfix/sdmem_on_intel_gpu branch (of the old Git repo), this can help for proper
graphics hardware reinitialization post-kexec.

We include:

 * bochs, cirrus and qxl because they can be useful in virtual machines;
 * i915, nouveau and radeon are they are the best maintained ones these days,
   and cover the vast majority of hardware that is able to run Tails/Jessie
   properly.