Revert "Switch to the first console as a first step when stopping the GDM service."
This reverts commit e297e6388f300d15efa8330069eb61118a7fbd50.

That was a nice addition, but it currently fully breaks the shutdown process,
and hence our "erase memory on shutdown" feature (except for the emergency
shutdown that is not affected): in non-quiet and nosplash mode, I've seen
a message similar to "a stop job for gdm.service is running", and this stop job
can't be anything else than the `/bin/chvt 1' we've added.

So let's drop this call, and see what happens.