Mask the plymouth-{halt,kexec,poweroff,reboot,shutdown} services.
This prevents it from occupying the active TTY with an (empty) splash screen
on shutdown/reboot, which hides the messages we want to show to the user
in /lib/systemd/system-shutdown/tails-kexec.

I've tried to actually use plymouth to display these messages, but I failed.
Anyway, now the messages are the only one displayed, and the scary long list of
"shutting down $service" messages is hidden, which is a pretty good UX
improvement already.

Closes: #9032