Start adjusting waiting for notification facilities for Jessie.
It makes no sense to wait for gnome-panel on Jessie: we're not running
it anymore. Also, GNOME Shell is now providing the notification daemon
facility, so we don't need to wait for notification-daemon.

Will-fix: #8685 (Well, not really: as said there, this might not be robust
enough: best would be to check if the DBus service is ready instead.)