Wait for ibus-daemon instead of nm-applet, to judge whether we can send notifications.
We should not be running nm-applet on GNOME Shell, since this functionality is
nowadays taken over by GNOME Shell itself. And notifications are also handled by
GNOME Shell, so let's instead wait for a process that's started by GNOME Shell,
assuming that if it's ready enough to start long-lived child processes,
hopefully it's ready to receive and enqueue/display notifications.

Will-fix: #8685 (not really, but let's help Redmine track the relevant commits)