Make Desktop::Notify support "hints".
Patch submitted upstream: https://rt.cpan.org/Ticket/Display.html?id=102644

This is required so that we can flag some notifications as "transient", which is
sometimes needed: GNOME Shell ignores the timeout we pass it. So, the
NotificationClosed signal is never send, and then the program that's waiting for
it would stay around forever, unless the user explicitly dismisses the
notification. This can lead to increased memory consumption. So, let's give us
the power to decide, for each notification we send, whether:

 * it should stay around until the user dismisses it (or clicks on the action
   button, if any), at the cost of some memory usage for the user, or of a more
   complicated system that's left to be invented and implemented by us;

or,

 * it will disappear after 4 seconds (see NOTIFICATION_TIMEOUT in
   js/ui/messageTray.js, in the GNOME Shell source code), regardless of whether
   the user has interacted with it.