Migrate to python-potr.
python-otr is not in Jessie, but python-potr is, and it's in
wheezy-backports.

This migration works as long as one converts the OTR key using the
convertkey.py tool that should be included in the debian package (even
according to its own description) but isn't. Currently it's in the
source package.