Fix test suite's TailsBootSplash* images for Jessie's virt-viewer.
Jessie's virt-viewer shows the display in fullscreen from start, not
only after resetting the VM, like in Wheezy.

Also, the only reason "Testing of reboot with UEFI enabled" wasn't
supported was that I didn't bother adding an image for it, as would
have been required with Wheezy's virt-viewer. Now when the same image
as when booting with UEFI enabled can be used we get this support for
free.