Rename --temp-dir to --tmpdir and make it behave more like mktemp.
That is, if --tmpdir is used, it takes precedence, otherwise we use
TMPDIR from the environment, and lastly, fallback to
/tmp/TailsToaster.

Also, apparently we've had a discrepancy where we used TEMP_DIR in
run_test_suite vs TMP_DIR everywhere else, so the old --temp-dir
option hasn't been working any way. With --tmpdir/TMPDIR this is now
consistent everywhere.

Will-fix: #8709