Re-do detached signature on the 1.3/i386/stable update-description file, with our new signing key.
We've merged feature/8740-new-signing-key-phase-2 into stable a few days ago, so
the code in stable/devel/etc. only trusts the new signing key, while the 1.3 UDF
are still signed with our old signing key. Tails 1.3 trusts both the old and new
signing keys, so this should be alright both for it and for anything that has
feature/8740-new-signing-key-phase-2 merged in.

Will-fix: #9085