Wrap syndaemon to always use -t (Will-fix: #9011).
Without the -t option, which makes syndaemon only disable tapping and
scrolling and not mouse movements, florence is unusable when syndaemon
is running. In Wheezy, GNOME invokes syndaemon without -t, and the
option it passes are hardcoded, so we're forced to do this ugly
workaround, which luckily can be removed once we're based on Jessie
since -t is passed then. For details, see #9011.