Create a tails-wait-until-tor-has-bootstrapped.service in the systemd user instance.
The systemd user and system instances don't share units. Thus, the dependency
we have from the security check and upgrader user services on
tails-wait-until-tor-has-bootstrapped had no effect.

So, we have the *system* tails-wait-until-tor-has-bootstrapped.service create
a file upon completion -- and the identically named *user* unit file wait for
that file to appear.