tails-custom-apt-sources: error out if config/APT_overlays contains duplicate lines.
It's probably a nice sanity check in general. We'll see how it goes in practice,
e.g. Jenkins builds are using the union merge driver, and thus this new check
may trigger build failures there. If/when this happens, then we can simply guard
the check with 'if [ -z "$JENKINS_URL" ]' so that it's ignored when building
from Jenkins => time will tell.