Replace Vidalia status icon by a GNOME shell extension
To get rid of unmaintained Vidalia, we need to replace its status icon. We add
a simple shell extension that adds a Tor status icon to the shell top bar. The
status is defined by the presence of a file that we create in our tor-is-ready
NetworkManager dispatcher script. This will-fix: #9002.