Allow throwing Timeout::Error in try_for() blocks (Will-fix: #9189).
That's quite useful, and it's pretty easily fixed with a new exception
class.

This fixes VMCommand.wait_until_remote_shell_is_up(), which violated
this precondition by doing exactly that (as part of its mechanism to
re-try checking that the remote shell is up). Now we hopefully
shouldn't see any remote shell hangups (on boot or after restoring a
snapshot) which plagued us when running the test suite on a Jessie
host in particular.