Test that the Tor Browser cannot access LAN resources.
Since no HTTP server is listening on the LAN resource the Tor Browser
would show the same "Unable to connect" error message even if the
request wasn't blocked. We should be fine, though, since we check the
captured traffic. The picture is more of a signal saying that "now the
network traffic would have been generated *if* the request isn't
blocked" which we use to know when to check the network capture.