Make the VM instance into a persistent global variable.
"Persistent" in the sense that it survives across scenarios. This is a
first step towards using a single Libvirt::Domain instance across the
full test suite run, which in turn is a step towards being able to use
Libvirt snapshots more easily (or at all, I think).

This was achieved with the following command:

    sed -i 's/@vm\>/$vm/g' -- features/step_definitions/*.rb \
                              features/support/*.rb \
                              features/support/helpers/*.rb