Work around Seahorse GUI glitchiness
When Seahorse appears to be frozen--apparently due to network issues--it
can often be worked around by refreshing the screen or activating a new
window. This makes selecting the key with the keyboard a bit more
difficult, hence moving to clicking the key in the list, then clicking
the `Import` button.

Once #9233 is fixed this problem may not be seen anymore. Still, it
may be good to do this to prevent graphical glitches from causing the
tests to fail.

Will-fix: #9343