Make sure the system clock isn't before the build date (Will-fix: #9149).
Our live-config hook that imports our signing keys depend on that the
system clock isn't before the date when the keys where created (sanity
check by gpg). To avoid this, we ensure (via an early live-config
hook) that the system clock cannot be before the Tails image build
date (up to one day, to deal with potential timezone
differences/mismatches).