Bypass shell for command.
Note the (last) `.join(' ')`, which makes all of this into a single
String and hence invoked by the shell.

Let's avoid using the shell unless it's necessary due to the potential
robustness and security issues that brings (e.g. plugging in some
variable that contain characters that has special meaning in the
shell).