Wait for systray icons to finish loading
When restoring snapshots in the test suite we'll wait for the for the
Florence icon to be displayed in the system tray to ensure that the
icons won't shift positions because another application icon was added
to the system tray.

Once the icon for Florence Virtual Keyboard has loaded, the only icon
left to load is Vidalia. Vidalia's loading will not cause the other
systray icons to shift positions.

Will-fix: #9258