Store complete URL in tbb-dist-url.txt.
Our installation script determines the Tor Browser version from the
tarball filenames and then looks in "$(cat
tbb-dist-url.txt)/$VERSION", which isn't very flexible. For instance,
for the pre-release of Tor Browser 4.5 our script looks in:

    http://people.torproject.org/~mikeperry/builds/4.5/

when they in fact were stored in

    http://people.torproject.org/~mikeperry/builds/4.5-build5/

So let's just store the full URL to the directory storing the
tarballs.