Hide access to the Add-ons manager in the Unsafe Browser.
It can still be accessed through about:addons, but it's currently
broken (#9307). Also, we do not want users to install add-ons in the
Unsafe Browser (it's outside of the indended use cases, and it may get
weird since restarting it will clear the chroot => all extensions are
lost -- add to this that some extensions only work well after
restarting the browser), so let's just hide it.

(This change is borrowed from the equivalent change in the I2P Browser
in commit b504ba67303587a9bece81b7efad81b7b04bbff5.)

Will-fix: #9305