Retry GNOME application menu actions when it glitches.
The GNOME application menus seem to have issues with clicks or
hovering actions not registering, and hence sometimes submenus are not
opened when they should, and sometimes clicks on the final application
shortcut are lost. There seems to be a correlation between this and
CPU load on the host running the test suite.

We workaround this by simply re-trying the last action when it seem to
fail.

Will-fix: #8928 hopefully.